extract out common mock response in test util file
